extract navigation comps
