+++ thoughtbubble-mobile/App.tsx
@@ -9,7 +9,7 @@ import { ActivityIndicator, StatusBar } from 'react-native';
 import { useSelector } from 'react-redux';
 import { Provider as ReduxProvider } from 'react-redux';
 import { Provider as PaperProvider } from 'react-native-paper';
-import { AppTabsNavigator } from './src/components/navigation/AppTabsNavigator';
+import { AppTabsNavigator } from './src/navigation/AppTabsNavigator';
 import { LoginScreen } from './src/screens/LoginScreen';
 import { checkForIdToken } from './src/utils/asyncStorage';
 import { storeUserAction } from './src/actions/storeUserAction';
